project_root/requirements.txt
Django
djangorestframework
python-dotenv
openai